Unify create-project form and draft edit page into one lazy-draft flow
Replaces components/ProjectForm.tsx and the standalone edit page with a
single ProjectEditor component that runs in two modes depending on
whether a project id exists yet. A project is now created lazily — on
an explicit "Save draft" click, or implicitly the first time a task is
added — rather than eagerly when the create screen opens.

- New mode: fields live in local state only, nothing hits the server.
- Draft/live mode: every field autosaves individually (title/description
  etc. on blur, selects/checkboxes/skills on change), matching how task
  fields already worked; inline field errors in both modes.
- /suggest and /admin/projects are now drafts-list pages with a "New
  Project" button into a dedicated /suggest/new or /admin/projects/new
  form page (volunteers with zero drafts skip straight to the form).
- New-mode screen also offers a one-click Submit/Publish (creates the
  draft then publishes immediately) and Delete (discards the form).
- A small fixed autosave indicator ("Saving…"/"Saved") in draft mode.

Updates e2e/actions/projects.ts and affected specs for the new flows.

Closes #231
